services/api/requirements.txt
weaviate-client
langchain
pandas
tabulate
openai
tiktoken
python-dotenv
pydantic
requests
urllib3<2